Add basic multi-site ssl support
HAProxy, if given a directory of ssl certs instead of a single SSL cert,
will correctly parse and use all certs in the specified directory.  This
allows us to do SNI to host multiple domains from the same IP address.

This is a pre-requisite for hosting ORE from both
www.openresearchexchange.com and questionnaire.patientslikeme.com.

The 'site' configuration is unchanged for backwards compatibility
reasons.  A new configuration "sites" was provided.  If sites is
provided, site is ignored regardless of it's absence or presence.