Allow PROXY protocol
TCP streams from ELBs make us lose the client IP.  By using the PROXY
protocol, we can restore this.  We want ports 80 and 443 to function
without PROXY though, so that direct connections from either stagedash
or for testing purposes continue to work.

Delegate ports 9080 and 9443 to be the ports used.